# backend/assets/types/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pontiac API",
        "description": "The JSON API behind the React UI. Routes, parameters and \u2014 for endpoints that return a response DTO \u2014 the response shapes.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/team/{id}/resend-invitation": {
            "post": {
                "summary": "A new invitation link, for someone who has not set their password yet.",
                "operationId": "post_api_admin_team_resend",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-[13-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TeamMemberOutput"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/team/{id}": {
            "delete": {
                "summary": "Disables an assistant: they are signed out at their next request. Nothing is deleted.",
                "operationId": "delete_api_admin_team_disable",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-[13-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TeamMemberOutput"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/team/{id}/enable": {
            "post": {
                "operationId": "post_api_admin_team_enable",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-[13-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TeamMemberOutput"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/platform/admins/{id}/resend-invitation": {
            "post": {
                "operationId": "post_api_platform_admin_resend",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-[13-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuperAdminOutput"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/platform/admins/{id}": {
            "delete": {
                "summary": "Never yourself: the platform always keeps the person doing this.",
                "operationId": "delete_api_platform_admin_disable",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-[13-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuperAdminOutput"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/platform/admins/{id}/enable": {
            "post": {
                "operationId": "post_api_platform_admin_enable",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-[13-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuperAdminOutput"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "TeamMemberOutput": {
                "required": [
                    "id",
                    "email",
                    "fullName",
                    "role",
                    "active",
                    "loginStatus"
                ],
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "fullName": {
                        "type": "string"
                    },
                    "role": {
                        "title": "\"owner\" or \"assistant\".",
                        "type": "string"
                    },
                    "active": {
                        "type": "boolean"
                    },
                    "loginStatus": {
                        "title": "\"active\" (has a password), \"invited\" (link sent, not used yet) or \"none\".",
                        "type": "string"
                    },
                    "lastSignInAt": {
                        "title": "ISO 8601; null until they first sign in.",
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "SuperAdminOutput": {
                "required": [
                    "id",
                    "email",
                    "fullName",
                    "active",
                    "loginStatus",
                    "you"
                ],
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "fullName": {
                        "type": "string"
                    },
                    "active": {
                        "type": "boolean"
                    },
                    "loginStatus": {
                        "title": "\"active\", \"invited\" or \"none\".",
                        "type": "string"
                    },
                    "lastSignInAt": {
                        "title": "ISO 8601; null until they first sign in.",
                        "type": "string",
                        "nullable": true
                    },
                    "you": {
                        "title": "The super admin reading the list: they cannot disable themselves.",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ]
}